refactor(website): Upgrade dependencies and use `StarlightPage` component
