+++ website/package.json
@@ -16,12 +16,11 @@
 		"textlint:fix": "textlint --fix './src/content/docs/ja/**/*.{md,mdx}'"
 	},
 	"devDependencies": {
-		"@astrojs/netlify": "^3.0.4",
-		"@astrojs/prism": "^2.1.2",
-		"@astrojs/react": "^2.3.2",
-		"@astrojs/rss": "^2.4.4",
-		"@astrojs/starlight": "^0.13.0",
-		"@astrojs/vercel": "^5.0.1",
+		"@astrojs/netlify": "^5.1.2",
+		"@astrojs/prism": "^3.0.0",
+		"@astrojs/react": "^3.0.10",
+		"@astrojs/rss": "^4.0.5",
+		"@astrojs/starlight": "^0.19.0",
 		"@biomejs/wasm-web": "link:../packages/@biomejs/wasm-web",
 		"@codemirror/lang-javascript": "6.1.8",
 		"@codemirror/lang-json": "6.0.1",
@@ -34,25 +33,25 @@
 		"@types/node": "^18.18.0",
 		"@types/react": "^18.2.23",
 		"@types/react-dom": "^18.2.8",
-		"@uiw/react-codemirror": "4.20.2",
+		"@uiw/react-codemirror": "4.21.21",
 		"@vitejs/plugin-react": "^4.2.1",
 		"autoprefixer": "^10.4.16",
 		"codemirror-lang-rome-ast": "0.0.6",
 		"fast-diff": "^1.3.0",
-		"hast-util-to-html": "^8.0.4",
-		"hastscript": "^8.0.0",
+		"hast-util-to-html": "^9.0.0",
+		"hast-util-to-string": "^3.0.0",
+		"hastscript": "^9.0.0",
 		"html-escaper": "^3.0.3",
 		"lang-rome-formatter-ir": "0.0.2",
-		"mdast-util-to-hast": "^12.3.0",
+		"mdast-util-to-hast": "^13.1.0",
 		"mermaid": "^9.4.3",
 		"postcss": "^8.4.30",
 		"prettier": "^3.1.0",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"rehype-autolink-headings": "^6.1.1",
-		"rehype-slug": "^5.1.0",
-		"rehype-toc": "^3.0.2",
-		"remark-toc": "^8.0.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"rehype-autolink-headings": "^7.1.0",
+		"rehype-slug": "^6.0.0",
+		"remark-toc": "^9.0.0",
 		"rollup": "^3.29.4",
 		"sass": "^1.68.0",
 		"sharp": "^0.32.6",
@@ -67,7 +66,7 @@
 		"pnpm": "^8.0.0"
 	},
 	"dependencies": {
-		"astro": "^3.6.0",
+		"astro": "^4.4.0",
 		"astro-og-canvas": "^0.4.2",
 		"canvaskit-wasm": "^0.39.1"
 	}
