--- website/src/layouts/StarlightSplashLayout.astro
@@ -1,129 +1,0 @@
----
-// @ts-expect-error
-import type { LocaleData } from "@astrojs/starlight/utils/slugs";
-import Page from "@astrojs/starlight/components/Page.astro";
-
-export interface Props {
-	title: string;
-	description: string;
-	slug: string;
-	head?: Array<{
-		tag: string;
-		attrs?: Record<string, string | boolean | undefined>;
-		content?: string;
-	}>;
-	editUrl?: string;
-}
-
-function getPageProps({ title, description, slug, head = [], editUrl }: Props) {
-	const entryMeta = { dir: "ltr", lang: "en", locale: undefined } as LocaleData;
-
-	return {
-		...entryMeta,
-		editUrl: editUrl ? new URL(editUrl) : undefined,
-		entry: {
-			data: {
-				head: [
-					{
-						tag: "title",
-						content: `${title} | Biome`,
-					},
-					...head,
-				],
-				pagefind: false,
-				title,
-				description,
-				template: "splash",
-			},
-			slug,
-		},
-		entryMeta,
-		hasSidebar: true,
-		headings: [],
-		id: slug,
-		lastUpdated: undefined,
-		pagination: {
-			next: undefined,
-			prev: undefined,
-		},
-		sidebar: [
-			{
-				type: "link",
-				label: "Docs",
-				href: "/guides/getting-started",
-				isCurrent: Astro.url.pathname.includes("/guides/getting-started"),
-				attrs: {
-					class: "",
-				},
-			},
-			{
-				type: "link",
-				label: "Blog",
-				href: "/blog",
-				isCurrent: Astro.url.pathname.includes("/blog"),
-				attrs: {
-					class: "",
-				},
-			},
-			{
-				type: "link",
-				label: "Playground",
-				href: "/playground",
-				isCurrent: Astro.url.pathname.includes("/playground"),
-				attrs: {
-					class: "",
-				},
-			},
-		],
-		slug,
-		toc: {
-			items: [],
-			maxHeadingLevel: 0,
-			minHeadingLevel: 0,
-		},
-	};
-}
-
-const pageProps = getPageProps(Astro.props);
----
-
-<Page {...pageProps}>
-  <slot />
-</Page>
-
-<style>
-  :global(:root) {
-    --sl-content-width: 67.5rem;
-    --sl-content-inline-start: 0rem;
-  }
-
-  :global(.right-sidebar-container) {
-    display: none;
-  }
-
-  :global(.mobile-starlight-toc) {
-    display: none;
-  }
-
-  :global(.main-frame) {
-    padding-top: var(--sl-nav-height);
-  }
-
-  :global(.main-pane) {
-    width: 100% !important;
-  }
-
-  :global(.sl-container) {
-    margin-inline: auto;
-  }
-
-  :global(starlight-lang-select) {
-    display: none;
-  }
-
-  :global(@media (min-width: 50rem)) {
-    #starlight__sidebar {
-      display: none;
-    }
-  }
-</style>
