+++ website/src/pages/blog/[...slug].astro
@@ -1,19 +1,19 @@
 ---
 import { getCollection, type CollectionEntry } from "astro:content";
 import BlogLayout from "@/layouts/Blog.astro";
-import StarlightSplashLayout from "@/layouts/StarlightSplashLayout.astro";
+import StarlightPage from "@astrojs/starlight/components/StarlightPage.astro";
 import { getImage } from "astro:assets";
 
 export async function getStaticPaths() {
-	const blogEntries = await getCollection("blog");
-	return blogEntries.map((entry) => ({
-		params: { slug: entry.slug },
-		props: { entry },
-	}));
+  const blogEntries = await getCollection("blog");
+  return blogEntries.map((entry) => ({
+    params: { slug: entry.slug },
+    props: { entry },
+  }));
 }
 
 interface Props {
-	entry: CollectionEntry<"blog">;
+  entry: CollectionEntry<"blog">;
 }
 
 const { entry } = Astro.props;
@@ -23,36 +23,39 @@ const post = entry.data;
 
 const processedImage = await getImage({ src: post.socialImage });
 const socialImageUrl = new URL(
-	processedImage.src,
-	`${Astro.url.origin}/`,
+  processedImage.src,
+  `${Astro.url.origin}/`
 ).toString();
 
 const head = [
-	{
-		tag: "meta",
-		attrs: {
-			property: "og:image",
-			content: socialImageUrl,
-		},
-	},
-	{
-		tag: "meta",
-		attrs: {
-			property: "twitter:image",
-			content: socialImageUrl,
-		},
-	},
+  {
+    tag: "meta",
+    attrs: {
+      property: "og:image",
+      content: socialImageUrl,
+    },
+  },
+  {
+    tag: "meta",
+    attrs: {
+      property: "twitter:image",
+      content: socialImageUrl,
+    },
+  },
 ];
 ---
 
-<StarlightSplashLayout
-  title={post.title}
-  description={post.description}
-  slug={entry.slug}
-  head={head}
-  editUrl={`https://github.com/biomejs/biome/edit/main/website/src/content/blog/${entry.id}`}
+<StarlightPage
+  frontmatter={{
+    title: post.title,
+    description: post.description,
+    slug: entry.slug,
+    head: head,
+    template: "splash",
+    editUrl: `https://github.com/biomejs/biome/edit/main/website/src/content/blog/${entry.id}`,
+  }}
 >
   <BlogLayout post={post}>
     <Content />
   </BlogLayout>
-</StarlightSplashLayout>
+</StarlightPage>
